Test a range of stages in CircularBufferingTest (#2690)
This PR adds a parameterized test suite for circular buffering. It is
used to test a range of buffers from 2 to 4.

TODO: Add a runtime check to ensure that entire pipeline is filled by
given inputs. See #2699.

+++ csrc/device_lower/pass/circular_buffer.cpp
@@ -86,14 +86,12 @@ class CircularBufferLoopCloner : public kir::IrVisitor {
         requireEpilogue(circular_buffer_load_exprs_)) {
       stop = IrBuilder::subExpr(
           circular_buffer_loop_->stop(),
-          SimplifyingIrBuilder::create<Val>(
-              int64_t(stage_depth - 1), DataType::Index));
+          SimplifyingIrBuilder::create<Val>(stage_depth - 1, DataType::Index));
     } else if (loop_type_ == CircularBufferLoopStage::Epilog) {
       NVF_ERROR(requireEpilogue(circular_buffer_load_exprs_));
       start = IrBuilder::subExpr(
           circular_buffer_loop_->stop(),
-          SimplifyingIrBuilder::create<Val>(
-              int64_t(stage_depth - 1), DataType::Index));
+          SimplifyingIrBuilder::create<Val>(stage_depth - 1, DataType::Index));
     }
 
     cloned_top_level_loop_ = IrBuilder::create<ForLoop>(
